Add and use RenderableIntel
