+++ src/main/java/stelnet/BaseBoard.java
@@ -1,56 +1,8 @@
 package stelnet;
 
-import java.util.List;
+import stelnet.ui.RenderableIntel;
 
-import com.fs.starfarer.api.impl.campaign.intel.BaseIntelPlugin;
-import com.fs.starfarer.api.ui.CustomPanelAPI;
-import com.fs.starfarer.api.ui.IntelUIAPI;
-import com.fs.starfarer.api.ui.TooltipMakerAPI;
-
-import stelnet.helper.LogHelper;
-import stelnet.ui.ButtonHandler;
-import stelnet.ui.RenderableView;
-import stelnet.ui.Size;
-
-public abstract class BaseBoard extends BaseIntelPlugin {
-
-    @Override
-    public void buttonPressCancelled(Object buttonId, IntelUIAPI ui) {
-        LogHelper.debug("Calling cancel()");
-        ButtonHandler handler = (ButtonHandler) buttonId;
-        handler.onCancel(ui);
-        redraw(ui);
-    }
-
-    @Override
-    public void buttonPressConfirmed(Object buttonId, IntelUIAPI ui) {
-        LogHelper.debug("Calling confirm()");
-        ButtonHandler handler = (ButtonHandler) buttonId;
-        handler.onConfirm(ui);
-        redraw(ui);
-    }
-
-    @Override
-    public void createConfirmationPrompt(Object buttonId, TooltipMakerAPI prompt) {
-        LogHelper.debug("Calling prompt()");
-        ButtonHandler handler = (ButtonHandler) buttonId;
-        handler.onPrompt(prompt);
-    }
-
-    @Override
-    public void createLargeDescription(CustomPanelAPI panel, float width, float height) {
-        Size size = new Size(width, height);
-        for (RenderableView view : getRenderableViews()) {
-            view.render(panel, size);
-        }
-    }
-
-    @Override
-    public boolean doesButtonHaveConfirmDialog(Object buttonId) {
-        LogHelper.debug("Calling hasPrompt()");
-        ButtonHandler handler = (ButtonHandler) buttonId;
-        return handler.hasPrompt();
-    }
+public class BaseBoard extends RenderableIntel {
 
     @Override
     public boolean hasLargeDescription() {
@@ -58,19 +10,7 @@ public abstract class BaseBoard extends BaseIntelPlugin {
     }
 
     @Override
-    public boolean hasSmallDescription() {
-        return false;
-    }
-
-    @Override
     public IntelSortTier getSortTier() {
         return IntelSortTier.TIER_0;
-    }
-
-    protected abstract List<RenderableView> getRenderableViews();
-
-    private void redraw(IntelUIAPI ui) {
-        ui.recreateIntelUI();
-        ui.updateUIForItem(this);
     }
 }
